+++ composeApp/src/commonMain/kotlin/com/fruex/beerwall/domain/model/Transaction.kt
@@ -15,5 +15,5 @@ data class Transaction(
     val commodityName: String,
     val startDateTime: String,
     val grossPrice: Double,
-    val capacity: Int
+    val capacity: Double
 )
